+++ arctic_training/deepspeed.py
@@ -1118,101 +1118,6 @@ class UlyssesSPFwdLossBwdWithLogits:
         self.model.backward(self.loss)
 
 
-class TiledLoss(torch.autograd.Function):
-    @staticmethod
-    def forward(ctx, loss_fn, logits, vocab_size, shift_labels, shards) -> torch.Tensor:
-        """
-        logits doesn't have to be divisible by shards, the last shard will be shorter than the rest.
-        """
-        ctx.save_for_backward(logits, shift_labels)
-        ctx.loss_fn = loss_fn
-        ctx.vocab_size = vocab_size
-        ctx.shards = shards
-
-        with torch.no_grad():
-            seqlen = shift_labels.shape[1]
-            shard_step = math.ceil(seqlen / shards)
-            loss_shards = []
-            total_good_items = 0
-
-            # since -100s are ignored we have to perform a weighted average on each loss slice as each slice may contribute a different number of non- -100 labels
-            # if seqlen / shards != 0 - the last chunk is just shorter than the rest but no data is ignored
-            for i in range(shards):
-                # XXX: here and everywhere don't make a copy, pass the slice or perhaps narrow/view?
-                shift_labels_shard = shift_labels[:, i * shard_step : (i + 1) * shard_step]
-                if all((shift_labels_shard == -100).squeeze()):
-                    continue  # ignore this shard
-                loss_shard = loss_fn(
-                    logits=logits[:, i * shard_step : (i + 1) * shard_step, :],
-                    labels=None,
-                    vocab_size=vocab_size,
-                    shift_labels=shift_labels_shard,
-                )
-                good_items = sum((shift_labels_shard != -100).squeeze())
-                loss_shards.append(loss_shard * good_items)
-                total_good_items += good_items
-            total_loss = torch.cat([l.unsqueeze(0) for l in loss_shards], dim=0).sum()
-            weighted_loss = total_loss / total_good_items
-
-        # weighted_loss.requires_grad = True
-        return weighted_loss
-
-    @staticmethod
-    def backward(ctx, *grads) -> torch.Tensor:
-
-        logits, shift_labels = ctx.saved_tensors
-        loss_fn = ctx.loss_fn
-        vocab_size = ctx.vocab_size
-        shards = ctx.shards
-
-        grad = grads[0]
-        logits_grad = torch.zeros_like(logits)
-        # logits_grad = torch.zeros(logits.shape, device=logits.device, dtype=grad.dtype, requires_grad=logits.requires_grad)
-
-        logits_shards = list(torch.chunk(logits, chunks=shards, dim=1))
-        shift_labels_shards = list(torch.chunk(shift_labels, chunks=shards, dim=1))
-        del logits
-        del shift_labels
-        ctx.logits = None
-        ctx.shift_labels = None
-        ctx.loss_fn = None
-        ctx.vocab_size = None
-        ctx.shards = None
-
-        # if seqlen is not exactly divisible by shards the last step will be shorter than shard_step
-        shard_step = logits_shards[0].numel()
-        for i in range(shards):
-            logits_shard = logits_shards.pop(0)
-            shift_labels_shard = shift_labels_shards.pop(0)
-
-            shard_offset = i * shard_step
-            # this will enable gradual population of the pre-allocated `logits_shard.grad` during `torch.autograd.backward` calls
-            logits_shard.grad = (
-                logits_grad.view(-1).narrow(0, shard_offset, logits_shard.numel()).view_as(logits_shard)
-            )
-
-            with torch.enable_grad():
-                if all((shift_labels_shard == -100).squeeze()):
-                    # fake loss calculation, since CE will return nan, but grads will be set
-                    # a normal loss_fn upcasts logits to float so match it
-                    loss_shard = (logits_shard.sum() * 0.0).float()
-                else:
-                    loss_shard = loss_fn(
-                        logits=logits_shard.requires_grad_(),
-                        labels=None,
-                        vocab_size=vocab_size,
-                        shift_labels=shift_labels_shard,
-                    )
-
-            torch.autograd.backward(loss_shard, grad)
-
-        logits_grad /= shards
-
-        # print(f"returning {logits_grad.norm()=}")
-        # print(f"returning {logits_grad=}")
-        # only logits (2nd arg) needs grads
-        return None, logits_grad, None, None, None
-
 
 def sequence_tiled_compute(
     fn,
